98 use double for numbers (#101)
* [perf] change decimals to real/float in entities

* [perf] change double to real/float in dtos and funcs

* [fix] change coordinate back to double

+++ Backend/Entities/BrewGrinderDialSetting.cs
@@ -2,7 +2,7 @@ namespace Backend.Entities;
 
 public class BrewGrinderDialSetting : AuditableEntity
 {
-  public decimal DialValue { get; set; }
+  public float DialValue { get; set; }
 
   public int BrewSettingId { get; set; }
   public int GrinderDialId { get; set; }
